Use shared callRealApi function
