+++ packages/gatsby/src/services/run-mutation-batch.ts
@@ -1,20 +1,5 @@
-import { IMutationAction } from "../state-machines/data-layer/types"
-import { Store, AnyAction } from "redux"
-import { IGatsbyState } from "../redux/types"
 import { IWaitingContext } from "../state-machines/waiting/types"
-import { assertStore } from "../utils/assert-store"
-import { actions } from "../redux/actions"
-
-const callRealApi = (
-  event: IMutationAction,
-  store?: Store<IGatsbyState, AnyAction>
-): void => {
-  assertStore(store)
-  const { type, payload } = event
-  if (type in actions) {
-    store.dispatch(actions[type](...payload))
-  }
-}
+import { callRealApi } from "../utils/call-deferred-api"
 
 // Consume the entire batch and run actions
 export const runMutationBatch = async ({
