+++ packages/gatsby/src/state-machines/develop/actions.ts
@@ -6,39 +6,15 @@ import {
   ActionFunctionMap,
   DoneEventObject,
 } from "xstate"
-import { Store } from "redux"
-import { IBuildContext, IMutationAction } from "../../services"
-import { actions, boundActionCreators } from "../../redux/actions"
+import { IBuildContext } from "../../services"
+import { boundActionCreators } from "../../redux/actions"
 import { listenForMutations } from "../../services/listen-for-mutations"
 import { DataLayerResult } from "../data-layer"
-import { assertStore } from "../../utils/assert-store"
 import { saveState } from "../../db"
 import reporter from "gatsby-cli/lib/reporter"
 import { ProgramStatus } from "../../redux/types"
 import { createWebpackWatcher } from "../../services/listen-to-webpack"
-
-/**
- * These are the deferred redux actions sent from api-runner-node
- * They may include a `resolve` prop (if they are createNode actions).
- * If so, we resolve the promise when we're done
- */
-export const callRealApi = (event: IMutationAction, store?: Store): void => {
-  assertStore(store)
-  const { type, payload, resolve } = event
-  if (type in actions) {
-    // If this is a createNode action then this will be a thunk.
-    // No worries, we just dispatch it like any other
-    const action = actions[type](...payload)
-    const result = store.dispatch(action)
-    // Somebody may be waiting for this
-    if (resolve) {
-      resolve(result)
-    }
-  } else {
-    reporter.log(`Could not dispatch unknown action "${type}`)
-  }
-}
-
+import { callRealApi } from "../../utils/call-deferred-api"
 /**
  * Handler for when we're inside handlers that should be able to mutate nodes
  * Instead of queueing, we call it right away
